+++ src/components/popup/title-bar.vue
@@ -67,6 +67,7 @@
 <script>
 import { computed } from 'vue'
 import VIcon from '../icon/index.vue'
+import usePopupTitleBarBase from '../../composables/usePopupTitleBarBase'
 
 export default {
   name: 'v-popup-title-bar',
@@ -76,41 +77,8 @@ export default {
   },
 
   props: {
-    // The title of popup title bar
-    title: {
-      type: String,
-      default: ''
-    },
-    // The description of popup title bar
-    describe: {
-      type: String,
-      default: ''
-    },
-    // The confirmation text
-    okText: {
-      type: String,
-      default: ''
-    },
-    // The cancellation text
-    cancelText: {
-      type: String,
-      default: ''
-    },
-    // The position of title and description
-    titleAlign: {
-      type: String,
-      default: 'center'
-    },
-    // Whether display large radius
-    largeRadius: {
-      type: Boolean,
-      default: false
-    },
-    // Whether display right close button only
-    onlyClose: {
-      type: Boolean,
-      default: false
-    }
+    // Merge base props
+    ...usePopupTitleBarBase()
   },
 
   setup(props, { emit }) {